requirements.txt
appdirs==1.4.4
avro-python3==1.10.2
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cramjam==2.3.2
dask==2021.7.0
fastavro==1.4.3
fastparquet==0.6.3
fsspec==2021.6.1
idna==2.10
importlib-metadata==4.6.1
locket==0.2.1
loguru==0.5.3
mypy-extensions==0.4.3
numpy==1.21.0
pandas==1.3.0
pandavro==1.6.0
partd==1.2.0
pathspec==0.8.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
six==1.16.0
thrift==0.13.0
toml==0.10.2
toolz==0.11.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
